+++ src/index.js
@@ -1,76 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-
-const Header = ({ name }) => {
-  return (
-    <h1>{name}</h1>
-  )
-}
-
-const Total = ({ parts }) => {
-  const sum = parts.reduce((sum, part) => {
-    return sum + part.exercises
-  }, 0)
-  return(
-    <p>Number of exercises {sum}</p>
-  ) 
-}
-
-const Part = ({part}) => {
-  return (
-    <p>
-      {part.name} {part.exercises}
-    </p>    
-  )
-}
-
-const Content = ({ parts }) => {
-  return (
-    <div>
-      {parts.map(part => <Part key={part.id} part={part} />)}
-    </div>
-  )
-}
-
-const Course = ({ course }) => {
-  return (
-    <>
-      <Header name={course.name} />
-      <Content parts={course.parts} />
-      <Total parts={course.parts} />
-    </>
-  )
-}
-
-const App = () => {
-  const course = {
-    id: 1,
-    name: 'Half Stack application development',
-    parts: [
-      {
-        name: 'Fundamentals of React',
-        exercises: 10,
-        id: 1
-      },
-      {
-        name: 'Using props to pass data',
-        exercises: 7,
-        id: 2
-      },
-      {
-        name: 'State of a component',
-        exercises: 14,
-        id: 3
-      },
-      {
-        name: 'Redux',
-        exercises: 11,
-        id: 4
-      }
-    ]
-  }
-
-  return <Course course={course} />
-}
+import App from './App'
 
 ReactDOM.render(<App />, document.getElementById('root'))
